+++ src/main/java/com/buuz135/industrial/proxy/CommonProxy.java
@@ -61,7 +61,7 @@ public class CommonProxy {
         RecipeHandlers.loadBioReactorEntries();
         RecipeHandlers.loadLaserLensEntries();
         RecipeHandlers.loadSludgeRefinerEntries();
-
+        RecipeHandlers.loadProteinReactorEntries();
     }
 
     public void postInit() {