+++ src/main/java/com/buuz135/industrial/tile/generator/ProteinReactorTile.java
@@ -1,31 +1,31 @@
 package com.buuz135.industrial.tile.generator;
 
-import com.buuz135.industrial.api.recipe.BioReactorEntry;
 import com.buuz135.industrial.api.recipe.IReactorEntry;
+import com.buuz135.industrial.api.recipe.ProteinReactorEntry;
 import com.buuz135.industrial.proxy.BlockRegistry;
 import com.buuz135.industrial.proxy.FluidsRegistry;
 import net.minecraftforge.fluids.Fluid;
 
 import java.util.List;
 
-public class BioReactorTile extends AbstractReactorTile {
+public class ProteinReactorTile extends AbstractReactorTile {
 
-    public BioReactorTile() {
-        super(BioReactorTile.class.getName().hashCode());
+    public ProteinReactorTile() {
+        super(ProteinReactorTile.class.getName().hashCode());
     }
 
     @Override
     public List<IReactorEntry> getReactorsEntries() {
-        return BioReactorEntry.BIO_REACTOR_ENTRIES;
+        return ProteinReactorEntry.PROTEIN_REACTOR_ENTRIES;
     }
 
     @Override
     public Fluid getProducedFluid() {
-        return FluidsRegistry.BIOFUEL;
+        return FluidsRegistry.PROTEIN;
     }
 
     @Override
     public int amountProduced() {
-        return BlockRegistry.bioReactorBlock.getBaseAmount();
+        return BlockRegistry.proteinReactorBlock.getBaseAmount();
     }
 }